Add audio-player module
Contains the composable which hosts it, along with the card, and
contains all the internal logic to handle loading, error states, and
dragging on the component itself to listen to specific parts of the
audio content
