+++ app/build.gradle.kts
@@ -118,6 +118,7 @@ dependencies {
   implementation(projects.apollo.core)
   implementation(projects.apollo.di)
   implementation(projects.apollo.giraffe)
+  implementation(projects.audioPlayer)
   implementation(projects.auth.authAndroid)
   implementation(projects.auth.authCore)
   implementation(projects.coreCommon)
