--- app/src/main/kotlin/com/hedvig/app/service/audioplayer/AudioPlayer.kt
@@ -1,15 +1,0 @@
-package com.hedvig.app.service.audioplayer
-
-import com.hedvig.app.util.ProgressPercentage
-import kotlinx.coroutines.flow.StateFlow
-import java.io.Closeable
-
-interface AudioPlayer : Closeable {
-  val audioPlayerState: StateFlow<AudioPlayerState>
-
-  fun initialize()
-  fun startPlayer()
-  fun pausePlayer()
-  fun retryLoadingAudio()
-  fun seekTo(progressPercentage: ProgressPercentage)
-}
