+++ audio-player/build.gradle.kts
@@ -1,7 +1,11 @@
 plugins {
+  id("hedvig.android.ktlint")
   id("hedvig.android.library")
   id("hedvig.android.library.compose")
-  id("hedvig.android.ktlint")
+}
+
+android {
+  namespace = "com.hedvig.android.audio.player"
 }
 
 dependencies {
@@ -9,21 +13,16 @@ dependencies {
   implementation(projects.coreDesignSystem)
   implementation(projects.coreResources)
 
-  api(libs.accompanist.insetsUi)
-  api(libs.androidx.compose.foundation)
-  api(libs.androidx.compose.material)
-  api(libs.androidx.compose.material3)
-  api(libs.arrow.core)
-
-  implementation(libs.androidx.compose.material3.windowSizeClass)
+  implementation(libs.androidx.compose.animation)
+  implementation(libs.androidx.compose.foundation)
+  implementation(libs.androidx.compose.material)
   implementation(libs.androidx.compose.materialIconsExtended)
   implementation(libs.androidx.compose.uiUtil)
+  implementation(libs.androidx.lifecycle.common)
+  implementation(libs.androidx.lifecycle.compose)
   implementation(libs.androidx.lifecycle.runtime)
-  implementation(libs.coil.coil)
-  implementation(libs.coil.compose)
   implementation(libs.slimber)
-}
 
-android {
-  namespace = "com.hedvig.android.core.ui"
+  testImplementation(libs.assertK)
+  testImplementation(libs.junit)
 }
