+++ feature-odyssey/src/main/kotlin/com/hedvig/android/odyssey/model/AudioUrl.kt
@@ -1,0 +1,7 @@
+package com.hedvig.android.odyssey.model
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+@JvmInline
+internal value class AudioUrl(val value: String)
